Adjust schema and seed database
Adjusted the database schema in several ways.
1. Added combo selection table to keep track of
which options were chosen for a combo in a given
sale
2. Adjusted the type of user pin to text, since
a leading 0 should be possible.
3. Added main_item_type and main_item_quantity to
combos to identify what item (or items in the
case of pizza for example) is the primary item
this will reference the product "type" which will
be a comma-separated list of keywords
(in addition to an assumed chip and drink option)
4. Added a combo_option_type column to products
which will make it easy to assign a product as a
valid chip or drink option in a combo (trying to
make this as data-driven as possible)