Add on close cleanup
If the program is closed, all shifts and the session are
automatically ended. I don't know if that's a good idea or not.
I probably want to consider on startup checking for open shifts
or sessions and automatically closing them. Or maybe something
more advanced like checking how long it's been since the start
and leaving things active in case of a crash?