Fix crash when clearing ticket
Attempting to clear the ticket when an item was selected
resulted in a crash. This was caused by the activeItem (which
was technically an index) pointing to a, now, non-existing index
of the sale.items array. By setting the activeItemIndex to null
first, before resetting the sale object, the issue is avoided.